+++ src/rtcModule/rtcStats.cpp
@@ -54,7 +54,7 @@ std::string Stats::getJson()
         std::vector<float> periods;
         for (unsigned int i = 1; i < mSamples.mT.size(); i++)
         {
-            periods.push_back(static_cast<float>(mSamples.mT[i] - mSamples.mT[i - 1])/1000.0);
+            periods.push_back(static_cast<float>(mSamples.mT[i] - mSamples.mT[i - 1]/1000.0));
         }
 
         rapidjson::Value t(rapidjson::kArrayType);
@@ -270,7 +270,7 @@ void ConnStatsCallBack::OnStatsDelivered(const rtc::scoped_refptr<const webrtc::
             mStats->mInitialTs = report->timestamp_us();
         }
 
-        mStats->mSamples.mT.push_back((report->timestamp_us() - mStats->mInitialTs)/ 1000);
+        mStats->mSamples.mT.push_back(static_cast<int32_t>((report->timestamp_us() - mStats->mInitialTs)/ 1000));
 
         for (auto it = report->begin(); it != report->end(); it++)
         {
@@ -281,10 +281,10 @@ void ConnStatsCallBack::OnStatsDelivered(const rtc::scoped_refptr<const webrtc::
                 int64_t bytesRecv = 0;
                 int64_t bytesSend = 0;
                 getConnStats(it, rtt, txBwe, bytesRecv, bytesSend);
-                mStats->mSamples.mRoundTripTime.back() += rtt;
-                mStats->mSamples.mOutGoingBitrate.back() += txBwe;
-                mStats->mSamples.mBytesReceived.back() += bytesRecv;
-                mStats->mSamples.mBytesSend.back() += bytesSend;
+                mStats->mSamples.mRoundTripTime.back() += rtt;       // note: upon update to GCC > 9 this warning should disappear
+                mStats->mSamples.mOutGoingBitrate.back() += txBwe;   // note: upon update to GCC > 9 this warning should disappear
+                mStats->mSamples.mBytesReceived.back() += bytesRecv; // note: upon update to GCC > 9 this warning should disappear
+                mStats->mSamples.mBytesSend.back() += bytesSend;     // note: upon update to GCC > 9 this warning should disappear
             }
             else if (strcmp(it->type(), "inbound-rtp") == 0)
             {
